% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTerra.R
\name{getTerra}
\alias{getTerra}
\title{getTerra
Query and load into R data from http://www.climatologylab.org/terraclimate.html}
\usage{
getTerra(start = 1958, end = 2017, byyear = 10, listyears = NULL,
  variableclim = "ppt", path = "data")
}
\arguments{
\item{start}{start}

\item{listyears}{Instead of start, end, and byyer, user can provide the list of year}

\item{path}{by default, "data"}
}
\value{
list of rasterstack
}
\description{
getTerra
Query and load into R data from http://www.climatologylab.org/terraclimate.html
}
